Use doubles everywhere, cleanup in /common/.
+ Add a physics engine submodule.
+ Use double precision everywhere.
+ Restructure things in libcommon.
+ Begin defining components for ECS.

+++ game/client/events/cursor_move.hpp
@@ -13,9 +13,9 @@
 namespace events
 {
 struct CursorMove final {
-    float xpos;
-    float ypos;
-    float dx;
-    float dy;
+    double xpos;
+    double ypos;
+    double dx;
+    double dy;
 };
 } // namespace events
